Add blocking, friending, and deleting

+++ controllers/usersController.js
@@ -1,6 +1,5 @@
 const db = require("../db/queries");
 const bcrypt = require("bcryptjs");
-const passport = require("passport");
 
 const keepParams = (req) => {
   const requestData = new URLSearchParams(req.query);
@@ -8,13 +7,16 @@ const keepParams = (req) => {
 };
 
 exports.displayGet = async (req, res) => {
-  const messages = (await db.getAllMessages(req.query)) || [];
-  console.log(req.session);
+  const messages =
+    (await db.getAllMessages(req.query, req.user?.id || 0)) || [];
+  const blocked = req.user ? await db.getBlockedUsers(req.user.id) : [];
+  console.log("blocked", blocked, req.user);
   res.render("index", {
     title: "Members Only Message Board",
     messages: messages || false,
-    authMessage: req.session.message,
+    authMessage: req.session?.messages?.at(-1) || req.session?.message || "",
     user: req.user || false,
+    blocked: blocked || [],
   });
 };
 
@@ -39,18 +41,59 @@ exports.signUp = async (req, res, next) => {
   }
 };
 
-// TODO: Update this to sign in
-exports.signIn = async (req, res, next) => {
-  try {
-    passport.authenticate("local", {
-      successRedirect: "/",
-      failureRedirect: "/",
-      failureMessage: true,
-      successMessage: true,
-    });
-  } catch (err) {
-    return next(err);
+exports.signOut = (req, res, next) => {
+  req.logout((err) => {
+    if (err) {
+      return next(err);
+    }
+    res.redirect("/");
+  });
+};
+
+// Moved to app.js
+// exports.signIn = async (req, res, next) => {
+//   try {
+//     passport.authenticate("local", {
+//       successRedirect: "/",
+//       failureRedirect: "/",
+//       failureMessage: true,
+//       successMessage: true,
+//     });
+//   } catch (err) {
+//     return next(err);
+//   }
+// };
+
+exports.blockUser = async (req, res) => {
+  if (+req.user?.id === +req.body.target) {
+    req.session.message = "You cannot block yourself!";
+    return res.redirect("/?" + keepParams(req));
   }
+  req.session.message = "Blocked!";
+  await db.blockUser(req.user.id, req.body.target);
+  res.redirect("/?" + keepParams(req));
+};
+
+exports.unblockUser = async (req, res) => {
+  req.session.message = "Unblocked!";
+  await db.unblockUser(req.user.id, req.body.target);
+  res.redirect("/?" + keepParams(req));
+};
+
+exports.friendUser = async (req, res) => {
+  if (+req.user?.id === +req.body.target) {
+    req.session.message = "You cannot friend yourself!";
+    return res.redirect("/?" + keepParams(req));
+  }
+  req.session.message = "Friended!";
+  await db.friendUser(req.user.id, req.body.target);
+  res.redirect("/?" + keepParams(req));
+};
+
+exports.defriendUser = async (req, res) => {
+  req.session.message = "Defriended!";
+  await db.defriendUser(req.user.id, req.body.target);
+  res.redirect("/?" + keepParams(req));
 };
 
 // for filtering by name
@@ -60,8 +103,13 @@ exports.displayPost = (req, res) => {
 };
 
 exports.deleteById = async (req, res) => {
-  const id = +req.body.id;
-  await db.deleteId(id);
+  if (req.user?.status === "admin" || req.user?.username === req.body.creator) {
+    const id = +req.body.target;
+    await db.deleteId(id);
+    req.session.message = "Deleted!";
+  } else {
+    req.session.message = "You do not have permission to delete this message!";
+  }
   res.redirect("/?" + keepParams(req));
 };
 
@@ -91,6 +139,18 @@ exports.getTitlesById = async (req, res) => {
 
 exports.newBookGet = (req, res) => {
   res.render("form", { title: "Form" });
+};
+
+exports.submitMessage = async (req, res) => {
+  const { title, message } = req.body;
+  console.log(req.user?.id ? req.user.id : 0);
+  if (!req.user) {
+    req.session.message = "You must be logged in to post a message!";
+    return res.redirect("/?" + keepParams(req));
+  }
+  const userid = req.user.id ? req.user.id : 0;
+  const id = await db.addMessage(title, message, userid);
+  res.redirect("/?" + keepParams(req));
 };
 
 exports.newBookPost = async (req, res) => {
